+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Object has different properties!
 - Support for parsing extra multiparts.
 - Support for History-Info header.
 - Support platforms without support for `globalThis`.
+- Support for explicitly accepting or rejecting messages.
 - Ability to remove message listeners.
 - Interface for fetching the remote SIP URI.
 ### Fixed